dpll: fix mode-supported parsing according to netlink spec
The DPLL netlink spec defines mode-supported as:
  type: u32
  enum: mode
  multi-attr: true

This means it's a simple u32 value, not a nested array. The previous
code incorrectly tried to parse it as a nested array using
mnl_attr_for_each_nested, which resulted in an empty array.

Changed to parse mode-supported as a single u32 value according to
the specification, fixing the mismatch with Python YNL CLI output.

Signed-off-by: Petr Oros <poros@redhat.com>