dpll: introduce DPLL_PARSE_U32 helper macro for parsing without nlh
Add DPLL_PARSE_U32 macro that parses a U32 argument into a variable
without adding it to a netlink message. This allows code reuse in cases
where we need to parse an argument but control when/if it's added to nlh.

Refactored DPLL_PARSE_ATTR_U32 to use DPLL_PARSE_U32 internally, and
updated all manual U32 parsing patterns to use the helper.

Changed patterns in:
- cmd_device_show: id parsing
- cmd_device_set: id parsing
- cmd_pin_show: id, device parsing
- cmd_pin_set: id parsing

Before:
  if (dpll_argv_match(dpll, "id")) {
      char *str = dpll_argv_next(dpll);
      if (!str) {
          pr_err("id requires an argument\n");
          return -EINVAL;
      }
      if (get_u32(&id, str, 0)) {
          pr_err("invalid id: %s\n", str);
          return -EINVAL;
      }
      has_id = true;
  }

After:
  if (dpll_argv_match(dpll, "id")) {
      DPLL_PARSE_U32(dpll, "id", &id);
      has_id = true;
  }

Benefits:
- 25 lines of code eliminated
- Consistent error messages across all U32 parsing
- Single source of truth for U32 argument parsing logic
- Easy to extend (can add more DPLL_PARSE_* variants)

Signed-off-by: Petr Oros <poros@redhat.com>